[MOVE> #12 <AttributeConvertFailedException> global 패키지로 이동
