+++ src/main/java/com/jobits/dsm/benecia/domain/recruitment/code/ProgrammingLanguageCode.java
@@ -1,6 +1,6 @@
 package com.jobits.dsm.benecia.domain.recruitment.code;
 
-import com.jobits.dsm.benecia.domain.recruitment.exception.AttributeConvertFailedException;
+import com.jobits.dsm.benecia.global.exception.AttributeConvertFailedException;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 
@@ -43,7 +43,7 @@ public enum ProgrammingLanguageCode {
 
     public static ProgrammingLanguageCode find(String dbData) {
         return Optional.of(map.get(dbData))
-                .orElseThrow(() -> new IllegalStateException("adf"));
+                .orElseThrow(AttributeConvertFailedException::new);
     }
 
     @Converter
