+++ src/main/java/com/jobits/dsm/benecia/domain/recruitment/code/RecruitmentReportingTimeCode.java
@@ -1,6 +1,6 @@
 package com.jobits.dsm.benecia.domain.recruitment.code;
 
-import com.jobits.dsm.benecia.domain.recruitment.exception.AttributeConvertFailedException;
+import com.jobits.dsm.benecia.global.exception.AttributeConvertFailedException;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
